ui/pipeline/requirements.txt
pymupdf
pillow
python-dotenv
openai
